# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nivora API
  version: 0.1.0-dev
  description: Early Nivora API for the DevOps delivery control plane. Phase 4.2 supports minimal shell PipelineRun execution, controlled Kubernetes YAML DeploymentRun foundations, artifact/release binding, GitOps/Argo CD foundations, resource inventory and rollback planning, release orchestration, DevSecOps policy gates, SecretRef/Credential metadata, local auth/RBAC, approval/change-window/notification foundations, multi-cloud inventory foundations, safe host deployment planning/noop execution, durable runner protocol foundations, backend visualization read models, and lightweight observability/operations endpoints.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/pipeline-runs:
    get:
      summary: List PipelineRuns in the in-memory Phase 1.5 runtime
      responses:
        "200":
          description: PipelineRun records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineRunRecord"
    post:
      summary: Create and run a minimal shell PipelineRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineDefinition"
      responses:
        "201":
          description: PipelineRun created and executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRunRecord"
        "400":
          description: Invalid pipeline definition
...
  /api/v1/releases:
    get:
      summary: List Releases in the in-memory Phase 2.2 runtime
      responses:
        "200":
          description: Release records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreateReleaseResponse"
    post:
      summary: Create a Release and bind ReleaseArtifacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReleaseRequest"
      responses:
        "201":
          description: Release created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateReleaseResponse"
        "400":
          description: Invalid release definition
...
  /api/v1/deployments:
    get:
      summary: List DeploymentRuns in the in-memory Phase 2.1 runtime
      responses:
        "200":
          description: DeploymentRun records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeploymentRunRecord"
    post:
      summary: Create and run a non-destructive YAML DeploymentRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentDefinition"
      responses:
        "201":
          description: DeploymentRun created and dry-run validation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentRunRecord"
        "400":
          description: Invalid deployment definition or manifest
...
components:
...
  schemas:
...
    ArtifactWarning:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        severity:
          type: string
    ArtifactInspection:
      type: object
      properties:
        reference:
          type: object
          properties:
            type:
              type: string
            original:
              type: string
            normalized:
              type: string
            registry:
              type: string
            repository:
              type: string
            name:
              type: string
            version:
              type: string
            tag:
              type: string
            digest:
              type: string
            scheme:
              type: string
            immutable:
              type: boolean
            isDigestPinned:
              type: boolean
            isLatest:
              type: boolean
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactWarning"
    ArtifactResolution:
      type: object
      properties:
        reference:
          type: object
        digest:
          type: string
        digestQualifiedReference:
          type: string
        mediaType:
          type: string
        resolved:
          type: boolean
        resolvedAt:
          type: string
          format: date-time
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactWarning"
...
    Artifact:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
        version:
          type: string
        reference:
          type: string
        digest:
          type: string
        registry:
          type: string
        repository:
          type: string
        mediaType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        metadata:
          type: object
          additionalProperties:
            type: string
    Release:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        applicationId:
          type: string
        environmentId:
          type: string
        sourcePipelineRunId:
          type: string
        commit:
          type: string
        status:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    ReleaseArtifact:
      type: object
      properties:
        id:
          type: string
        releaseId:
          type: string
        artifactId:
          type: string
        name:
          type: string
        type:
          type: string
        role:
          type: string
        required:
          type: boolean
        reference:
          type: string
        digest:
          type: string
        digestReference:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    CreateReleaseRequest:
      type: object
      required: [apiVersion, kind, metadata, spec]
      properties:
        apiVersion:
          type: string
        kind:
          type: string
          enum: [Release]
        metadata:
          type: object
          required: [name]
          properties:
            name:
              type: string
        spec:
          type: object
          required: [version, artifacts]
          properties:
            version:
              type: string
            application:
              type: string
            environment:
              type: string
            sourcePipelineRunId:
              type: string
            commit:
              type: string
            resolveDigest:
              type: boolean
            requireDigest:
              type: boolean
            artifacts:
              type: array
              items:
                type: object
                required: [name, type, reference]
                properties:
                  name:
                    type: string
                  type:
                    type: string
                  role:
                    type: string
                  required:
                    type: boolean
                  reference:
                    type: string
                  resolveDigest:
                    type: boolean
                  requireDigest:
                    type: boolean
                  metadata:
                    type: object
                    additionalProperties:
                      type: string
    CreateReleaseResponse:
      type: object
      properties:
        release:
          $ref: "#/components/schemas/Release"
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
        bindings:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseArtifact"
        inspections:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactInspection"
        resolutions:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactResolution"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactWarning"
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        audits:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
...
    SecurityFinding:
      type: object
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [Unknown, Low, Medium, High, Critical]
        category:
          type: string
          enum: [vulnerability, secret, misconfiguration, license, signature, sbom, policy]
        target:
          type: string
        title:
          type: string
        description:
          type: string
        packageName:
          type: string
        installedVersion:
          type: string
        fixedVersion:
          type: string
        reference:
          type: string
...
    GateDecision:
      type: string
      enum: [allow, deny, warn, require_approval]
...
    PipelineRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pipelineId:
          type: string
          format: uuid
        correlationId:
          type: string
        status:
          type: string
          enum: [Pending, Queued, Running, Paused, Succeeded, Failed, Canceled, Timeout]
        failureReason:
          type: string
    PipelineDefinition:
      type: object
      required: [apiVersion, kind, metadata, spec]
      properties:
        apiVersion:
          type: string
        kind:
          type: string
          enum: [Pipeline]
        metadata:
          type: object
          required: [name]
          properties:
            name:
              type: string
        spec:
          $ref: "#/components/schemas/PipelineSpec"
    PipelineSpec:
      type: object
      required: [stages]
      properties:
        timeoutSeconds:
          type: integer
          minimum: 0
        stages:
          type: array
          items:
            type: object
            required: [name, jobs]
            properties:
              name:
                type: string
              jobs:
                type: array
                items:
                  type: object
                  required: [name, executor, steps]
                  properties:
                    name:
                      type: string
                    executor:
                      type: string
                      enum: [shell]
                    retries:
                      type: integer
                      minimum: 0
                    timeoutSeconds:
                      type: integer
                      minimum: 0
                    steps:
                      type: array
                      items:
                        type: object
                        required: [run]
                        properties:
                          name:
                            type: string
                          run:
                            type: string
                          timeoutSeconds:
                            type: integer
                            minimum: 0
    PipelineRunRecord:
      type: object
      properties:
        pipeline:
          type: object
        run:
          $ref: "#/components/schemas/PipelineRun"
        stages:
          type: array
          items:
            $ref: "#/components/schemas/StageRun"
        logs:
          type: array
          items:
            $ref: "#/components/schemas/LogChunk"
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
    StageRun:
      type: object
      properties:
        stage:
          type: object
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/JobRun"
    JobRun:
      type: object
      properties:
        job:
          type: object
        steps:
          type: array
          items:
            $ref: "#/components/schemas/StepRun"
    StepRun:
      type: object
      properties:
        id:
          type: string
        jobRunId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [Pending, Assigned, Running, Succeeded, Failed, Skipped, Retrying, Canceled]
        attempt:
          type: integer
        failureReason:
          type: string
...
    LogChunk:
      type: object
      properties:
        id:
          type: string
        pipelineRunId:
          type: string
        deploymentRunId:
          type: string
        stageRunId:
          type: string
        jobRunId:
          type: string
        stepRunId:
          type: string
        stream:
          type: string
          enum: [stdout, stderr, system]
        sequence:
          type: integer
          format: int64
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    DeploymentRun:
      type: object
      properties:
        id:
          type: string
        releaseId:
          type: string
        applicationId:
          type: string
        correlationId:
          type: string
        environmentId:
          type: string
        releaseTargetId:
          type: string
        targetType:
          type: string
        status:
          type: string
          enum: [Created, Planning, WaitingApproval, PreChecking, Deploying, Verifying, Succeeded, Failed, RollingBack, RolledBack, Canceled]
        reason:
          type: string
        manifestSnapshotRef:
          type: string
        artifactReferences:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
    DeploymentDefinition:
      type: object
      required: [apiVersion, kind, metadata, spec]
      properties:
        apiVersion:
          type: string
        kind:
          type: string
          enum: [Deployment]
        metadata:
          type: object
          required: [name]
          properties:
            name:
              type: string
        spec:
          $ref: "#/components/schemas/DeploymentSpec"
    DeploymentSpec:
      type: object
      required: [application, environment, target]
      properties:
        application:
          type: string
        environment:
          type: string
        target:
          type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [kubernetes-yaml, argocd, host]
            name:
              type: string
            context:
              type: string
            namespace:
              type: string
            applicationName:
              type: string
            repoURL:
              type: string
            path:
              type: string
            revision:
              type: string
            project:
              type: string
            clusterURL:
              type: string
            clusterName:
              type: string
            syncPolicy:
              type: string
            credentialsRef:
              type: string
        artifact:
          $ref: "#/components/schemas/DeploymentArtifact"
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/DeploymentArtifact"
        manifests:
          type: array
          items:
            type: string
        gitops:
          type: object
          properties:
            mode:
              type: string
              enum: [plan]
            writeToWorkingTree:
              type: boolean
              default: false
            workingTree:
              type: string
            sync:
              type: boolean
              default: false
            statusRead:
              type: boolean
              default: false
            files:
              type: array
              items:
                type: string
        host:
          $ref: "#/components/schemas/HostDeploymentSpec"
        options:
          type: object
          properties:
            dryRun:
              type: boolean
            apply:
              type: boolean
            wait:
              type: boolean
            timeoutSeconds:
              type: integer
              minimum: 0
    DeploymentRunRecord:
      type: object
      properties:
        definition:
          $ref: "#/components/schemas/DeploymentDefinition"
        release:
          type: object
        artifacts:
          type: array
          items:
            type: object
        environment:
          type: object
        target:
          type: object
        run:
          $ref: "#/components/schemas/DeploymentRun"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/DeploymentStep"
        plan:
          $ref: "#/components/schemas/DeploymentPlan"
        gitopsPlan:
          $ref: "#/components/schemas/GitOpsChangePlan"
        gitopsDiff:
          $ref: "#/components/schemas/GitOpsDiff"
        hostPlan:
          $ref: "#/components/schemas/HostDeploymentPlan"
        hostDetails:
          type: array
          items:
            $ref: "#/components/schemas/HostDeploymentRunDetail"
        argocd:
          $ref: "#/components/schemas/ArgoCDApplicationStatus"
        inventory:
          $ref: "#/components/schemas/ResourceInventory"
        health:
          $ref: "#/components/schemas/HealthEvaluation"
        manifestSnapshot:
          $ref: "#/components/schemas/ManifestSnapshot"
        diff:
          $ref: "#/components/schemas/DeploymentDiff"
        rollbackPlan:
          $ref: "#/components/schemas/RollbackPlan"
        dryRun:
          $ref: "#/components/schemas/KubernetesDryRunResult"
        apply:
          $ref: "#/components/schemas/KubernetesApplyResult"
        rollout:
          $ref: "#/components/schemas/RolloutResult"
        rollback:
          $ref: "#/components/schemas/RollbackRecord"
        logs:
          type: array
          items:
            $ref: "#/components/schemas/LogChunk"
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        audits:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        policy:
          $ref: "#/components/schemas/PolicyResult"
    DeploymentPlan:
      type: object
      properties:
        deploymentRunId:
          type: string
        targetType:
          type: string
        namespace:
          type: string
        targetContext:
          type: string
        manifestCount:
          type: integer
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ManifestResourceSummary"
        artifacts:
          type: array
          items:
            type: string
        artifactDetails:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactInspection"
        manifestImages:
          type: array
          items:
            $ref: "#/components/schemas/ManifestImage"
        dryRun:
          type: boolean
        apply:
          type: boolean
        wait:
          type: boolean
        timeoutSeconds:
          type: integer
        actions:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        diffSummary:
          type: string
    DeploymentArtifact:
      type: object
      required: [name, type, reference]
      properties:
        name:
          type: string
        type:
          type: string
        reference:
          type: string
        digest:
          type: string
        target:
          type: object
          properties:
            imageName:
              type: string
            substitute:
              type: boolean
...
    HostDeploymentPlan:
      type: object
      properties:
        deploymentRunId:
          type: string
        groupName:
          type: string
        environmentId:
          type: string
        artifact:
          type: string
        deployPath:
          type: string
        serviceName:
          type: string
        healthCheck:
          type: string
        strategy:
          type: string
        dryRun:
          type: boolean
        apply:
          type: boolean
        hosts:
          type: array
          items:
            $ref: "#/components/schemas/HostDeploymentStep"
        actions:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        rollbackPlan:
          $ref: "#/components/schemas/RollbackPlan"
    HostDeploymentStep:
      type: object
      properties:
        hostId:
          type: string
        hostName:
          type: string
        address:
          type: string
        releaseDir:
          type: string
        currentSymlink:
          type: string
        previousSymlink:
          type: string
        nextSymlink:
          type: string
        actions:
          type: array
          items:
            type: string
    HostDeploymentRunDetail:
      type: object
      properties:
        hostId:
          type: string
        hostName:
          type: string
        address:
          type: string
        status:
          type: string
        message:
          type: string
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
    DeploymentStep:
      type: object
      properties:
        id:
          type: string
        deploymentRunId:
          type: string
        name:
          type: string
        status:
          type: string
        reason:
          type: string
...
    GitOpsChangePlan:
      type: object
      properties:
        deploymentRunId:
          type: string
        applicationName:
          type: string
        repoURL:
          type: string
        path:
          type: string
        revision:
          type: string
        files:
          type: array
          items:
            type: string
        fileChanges:
          type: array
          items:
            $ref: "#/components/schemas/GitOpsFileChange"
        artifactChanges:
          type: array
          items:
            type: string
        manifestValueChanges:
          type: array
          items:
            type: string
        commitMessageProposal:
          type: string
        dryRun:
          type: boolean
        warnings:
          type: array
          items:
            type: string
        syncRequested:
          type: boolean
        status:
          $ref: "#/components/schemas/ArgoCDApplicationStatus"
    GitOpsFileChange:
      type: object
      properties:
        path:
          type: string
        before:
          type: string
        after:
          type: string
        diff:
          type: string
        changed:
          type: boolean
        operation:
          type: string
        warning:
          type: string
    GitOpsDiff:
      type: object
      properties:
        summary:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/GitOpsFileChange"
    ArgoCDApplicationStatus:
      type: object
      properties:
        applicationName:
          type: string
        project:
          type: string
        syncStatus:
          type: string
          enum: [Unknown, Synced, OutOfSync]
        healthStatus:
          type: string
          enum: [Unknown, Healthy, Progressing, Degraded, Suspended, Missing]
        revision:
          type: string
        targetRevision:
          type: string
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ArgoCDResourceStatus"
        conditions:
          type: array
          items:
            type: string
        observedAt:
          type: string
        rawSummary:
          type: string
        message:
          type: string
        warnings:
          type: array
          items:
            type: string
    ArgoCDResourceStatus:
      type: object
      properties:
        group:
          type: string
        version:
          type: string
        kind:
          type: string
        namespace:
          type: string
        name:
          type: string
        status:
          type: string
        health:
          type: string
        syncStatus:
          type: string
        message:
          type: string
...
    ManifestResourceSummary:
      type: object
      properties:
        apiVersion:
          type: string
        group:
          type: string
        version:
          type: string
        kind:
          type: string
        name:
          type: string
        namespace:
          type: string
        sourceFile:
          type: string
        index:
          type: integer
        labels:
          type: object
          additionalProperties:
            type: string
        annotations:
          type: object
          additionalProperties:
            type: string
        desiredHash:
          type: string
        status:
          type: string
        health:
          $ref: "#/components/schemas/ResourceHealth"
    ResourceHealth:
      type: string
      enum: [Unknown, Progressing, Healthy, Degraded, Missing, Suspended, Unsupported]
    ResourceInventory:
      type: object
      properties:
        deploymentRunId:
          type: string
        desired:
          type: array
          items:
            $ref: "#/components/schemas/ManifestResourceSummary"
        applied:
          type: array
          items:
            $ref: "#/components/schemas/ManifestResourceSummary"
        warnings:
          type: array
          items:
            type: string
    HealthEvaluation:
      type: object
      properties:
        deploymentRunId:
          type: string
        status:
          $ref: "#/components/schemas/ResourceHealth"
        resourcesChecked:
          type: integer
        healthy:
          type: integer
        degraded:
          type: integer
        warnings:
          type: array
          items:
            type: string
        resources:
          type: array
          items:
            type: object
    ManifestSnapshot:
      type: object
      properties:
        id:
          type: string
        deploymentRunId:
          type: string
        contentHash:
          type: string
        documentCount:
          type: integer
        resourceCount:
          type: integer
        storageRef:
          type: string
        inlineContent:
          type: string
    DeploymentDiff:
      type: object
      properties:
        deploymentRunId:
          type: string
        addedResources:
          type: array
          items:
            type: string
        removedResources:
          type: array
          items:
            type: string
        changedResources:
          type: array
          items:
            type: string
        unchangedResources:
          type: array
          items:
            type: string
        unknownLiveState:
          type: array
          items:
            type: string
        summary:
          type: string
        warnings:
          type: array
          items:
            type: string
    RollbackPlan:
      type: object
      properties:
        deploymentRunId:
          type: string
        previousSnapshotId:
          type: string
        currentSnapshotId:
          type: string
        targetType:
          type: string
        targetName:
          type: string
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ManifestResourceSummary"
        strategy:
          type: string
        executable:
          type: boolean
        warnings:
          type: array
          items:
            type: string
    ManifestImage:
      type: object
      properties:
        resourceKind:
          type: string
        resourceName:
          type: string
        container:
          type: string
        image:
          type: string
...
    KubernetesDryRunResult:
      type: object
      properties:
        mode:
          type: string
        message:
          type: string
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ManifestResourceSummary"
        warnings:
          type: array
          items:
            type: string
        stdout:
          type: string
        stderr:
          type: string
    KubernetesApplyResult:
      type: object
      properties:
        mode:
          type: string
        message:
          type: string
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ManifestResourceSummary"
        warnings:
          type: array
          items:
            type: string
        stdout:
          type: string
        stderr:
          type: string
    RolloutResult:
      type: object
      properties:
        mode:
          type: string
        message:
          type: string
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ManifestResourceSummary"
        warnings:
          type: array
          items:
            type: string
        stdout:
          type: string
        stderr:
          type: string
    RollbackRecord:
      type: object
      properties:
        id:
          type: string
        deploymentRunId:
          type: string
        strategy:
          type: string
        status:
          type: string
        targetType:
          type: string
        targetName:
          type: string
        manifestSnapshotRef:
          type: string
        resourceRefs:
          type: array
          items:
            type: string
        reason:
          type: string
    PolicyResult:
      type: object
      properties:
        id:
          type: string
        policyId:
          type: string
        subjectType:
          type: string
        subjectId:
          type: string
        decision:
          $ref: "#/components/schemas/GateDecision"
        reason:
          type: string
        findings:
          type: array
          items:
            $ref: "#/components/schemas/SecurityFinding"
        evaluatedAt:
          type: string
          format: date-time
        allowed:
          type: boolean
        reasons:
          type: array
          items:
            type: string
...
    Event:
      type: object
      properties:
        specversion:
          type: string
        id:
          type: string
        type:
          type: string
        source:
          type: string
        subject:
          type: string
        time:
          type: string
          format: date-time
        datacontenttype:
          type: string
        data:
          type: object
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actor_id:
          type: string
          format: uuid
        action:
          type: string
        subject:
          type: string